Refactor normalize.js
