+++ .gitignore
@@ -1,5 +1,5 @@
 *.log
 /gatsby-node.js
 /fetch.js
-/processField.js
+/normalize.js
 node_modules
